Very small optimizations in g3dlite string functions

+++ dep/src/g3dlite/FileSystem.cpp
@@ -679,7 +679,7 @@ std::string FilePath::concat(const std::string& dirname, const std::string& file
 
 
 std::string FilePath::ext(const std::string& filename) {
-    int i = filename.rfind(".");
+    int i = (int)filename.rfind('.');
     if (i >= 0) {
         return filename.substr(i + 1, filename.length() - i);
     } else {
@@ -692,7 +692,7 @@ std::string FilePath::baseExt(const std::string& filename) {
     int i = findLastSlash(filename);
 
 #   ifdef G3D_WIN32
-        int j = filename.rfind(":");
+        int j = (int)filename.rfind(':');
         if ((i == -1) && (j >= 0)) {
             i = j;
         }
@@ -708,7 +708,7 @@ std::string FilePath::baseExt(const std::string& filename) {
 
 std::string FilePath::base(const std::string& path) {
     std::string filename = baseExt(path);
-    int i = filename.rfind(".");
+    int i = (int)filename.rfind('.');
     if (i == -1) {
         // No extension
         return filename;
@@ -722,7 +722,7 @@ std::string FilePath::parent(const std::string& path) {
     int i = findLastSlash(removeTrailingSlash(path));
 
 #   ifdef G3D_WIN32
-        int j = path.rfind(":");
+        int j = (int)path.rfind(':');
         if ((i == -1) && (j >= 0)) {
             i = j;
         }
